Move languages to constants file
